% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs-get-cert.r
\name{cs_get_cert}
\alias{cs_get_cert}
\title{Get Certificate Object}
\usage{
cs_get_cert(sha256, as = c("df", "pem", "der"),
  sslmate_api_key = Sys.getenv("SSLMATE_API_KEY"))
}
\arguments{
\item{sha256}{hex-encoded SHA-256 digest of the (pre-)certificate.}

\item{as}{determines the return type. \code{df} returns a data frame, \code{pem} returns
an X.509 cert in PEM format (\code{raw}), \code{der} returns a DER encoded X.509 cert
that is decoded with \code{\link[openssl:read_cert]{openssl::read_cert()}}.}

\item{sslmate_api_key}{(optional) See \code{Note}.}
}
\description{
Uses the \href{https://sslmate.com/certspotter/api}{sslmate Cert Spotter API} to
retrieve the certificate object for a given SHA256 hash.
}
\note{
Unauthenticated access to the API may be subject to rate limits.
Get an API key \href{https://sslmate.com/account/api_credentials?login=1}{here}.
}
