% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs-list-certs.r
\name{cs_list_certs}
\alias{cs_list_certs}
\title{List Certificates}
\usage{
cs_list_certs(domain, expired = FALSE, duplicate = FALSE,
  sslmate_api_key = Sys.getenv("SSLMATE_API_KEY"))
}
\arguments{
\item{domain}{Return certificates for the given domain and all sub-domains.
Also returns matching wildcard certificates. Must be at or below a
registered domain.}

\item{expired}{Include expired certificates? For \code{TRUE} you'll need an
\href{https://sslmate.com/account/api_credentials?login=1}{sslmate API key}
passed in to the function or stored in the \code{SSLMATE_API_KEY} environment
variable. Default: \code{FALSE}. When false, only the most recent (pre-)certificate
is returned for any given tbsCertificate. Default: \code{FALSE}.}

\item{duplicate}{Include duplicate certificates?}

\item{sslmate_api_key}{(optional). See \code{expired} and \code{Note}.}
}
\value{
data frame (tibble) with columns:
\itemize{
\item \code{type}: (character) cert or precert
\item \code{dns_names}: (list) DNS identifiers, from both the Subject CN and the DNS SANs
\item \code{sha256}: (character) The hex-encoded SHA-256 digest of the raw X.509 (pre-)certificate
\item \code{pubkey_sha256}: (character) The hex-encoded SHA-256 digest of the Subject Public Key Info
\item \code{issuer}: (character) The distinguished name of the certificate's issuer
\item \code{not_before}: (character) The not before date, in RFC3339 format (e.g. 2016-06-16T00:00:00-00:00)
\item \code{not_after}: (character) The not after date, in RFC3339 format (e.g. 2016-06-16T00:00:00-00:00)
\item \code{logs}: (list) A list of Certificate Transparency logs containing this (pre-)certificate
\item \code{data}: (character) The raw X.509 (pre-)certificate, encoded in base64
}
}
\description{
Uses the \href{https://sslmate.com/certspotter/api}{sslmate Cert Spotter API} to
list all CT logged certificates for a given domain.
}
\note{
Unauthenticated access to the API may be subject to rate limits.
}
