% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-entries.r
\name{get_entries}
\alias{get_entries}
\title{Retrieve Entries from Log}
\usage{
get_entries(log_server, start, end)
}
\arguments{
\item{log_server}{CTL log server host and prefix fragment (do not include
`https://` prefix. If one is present it will be removed).}

\item{start, end}{0-based index of first & last entry to retrieve, in decimal.}

\item{...}{other parameters passed on to `httr::GET`}
}
\description{
Retrieve Entries from Log
}
\examples{
ll <- read_log_list()
ctl <- ll$logs$url[2]
sth <- get_sth(ctl)
x <- get_entries(ctl, sth$tree_size-30, sth$tree_size-1)
}
\references{
<https://tools.ietf.org/html/rfc6962#section-4.6>
}
