% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sth.r
\name{get_sth}
\alias{get_sth}
\title{Retrieve Latest Signed Tree Head}
\usage{
get_sth(log_server, ...)
}
\arguments{
\item{log_server}{CTL log server host and prefix fragment (do not include
\code{https://} prefix. If one is present it will be removed).}

\item{...}{other parameters passed on to \code{httr::GET}}
}
\value{
data frame with five columns:
\itemize{
\item \code{log_server} what was passed in with any \code{https://} prefix stripped
\item \code{tree_Size} size of the tree, in entries, in decimal.
\item \code{timestamp} timestamp
\item \code{sha256_root_hash} Merkle Tree Hash of the tree, in base64.
\item \code{tree_head_signature} TreeHeadSignature for the above data (sans \code{log_server})
}
}
\description{
Retrieve Latest Signed Tree Head
}
\examples{
ll <- read_log_list()
ctl <- ll$logs$url[2]
sth <- get_sth(ctl)
x <- get_entries(ctl, sth$tree_size-30, sth$tree_size-1)
}
\references{
\url{https://tools.ietf.org/html/rfc6962#section-4.3}
}
