% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-log-list.r
\name{read_log_list}
\alias{read_log_list}
\title{Retrieve Certificate Transparency Log List}
\usage{

  read_log_list(log_list_url = "https://www.gstatic.com/ct/log_list/log_list.json")
}
\arguments{
\item{log_list_url}{a full URL to a known, good CTL log list. Defaults to a
known good one (at the time of package creation) hosted by Google.}
}
\value{
list
}
\description{
Retrieve Certificate Transparency Log List
}
\examples{
ll <- read_log_list()
ctl <- ll$logs$url[2]
sth <- get_sth(ctl)
x <- get_entries(ctl, sth$tree_size-30, sth$tree_size-1)
}
\references{
\url{https://www.certificate-transparency.org/known-logs}
}
