% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym-ct-search.r
\name{sym_ct_search}
\alias{sym_ct_search}
\title{Search Certificate Transparency Logs via Symatec CryptoReport}
\usage{
sym_ct_search(query)
}
\arguments{
\item{query}{domain, host, URL, organization name or serial number}
}
\value{
data frame or \code{NULL} (if no results)
}
\description{
Search Certificate Transparency Logs via Symatec CryptoReport
}
\note{
This performs a live search and the Symantec server can take up to or
longer than 30s to return.
}
\references{
\url{https://cryptoreport.websecurity.symantec.com/checker/views/ctsearch.jsp}
}
