% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr-summary.r
\name{tr_log_summary}
\alias{tr_log_summary}
\title{Retrieve Certificate Transparency Log Server Summaries from the
Google Transparency Report Project}
\usage{
tr_log_summary()
}
\value{
data frame (tibble)
}
\description{
Retrieve Certificate Transparency Log Server Summaries from the
Google Transparency Report Project
}
\references{
\url{https://transparencyreport.google.com/https/certificates}
}
