% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr-report.r
\name{tr_report}
\alias{tr_report}
\title{Query Google's Transparency Repoirt for Certificate Information}
\usage{
tr_report(hostname, include_subdomains = FALSE, include_expired = FALSE)
}
\arguments{
\item{hostname}{host name (domain)}

\item{include_subdomains, include_expired}{include subdomains and exipred certs?
Both default to \code{FALSE}}
}
\value{
list of 2 data frames; one for issuer metadata and one for host/cert metadata
}
\description{
TODO: "Drill down" for more info
}
