% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{parse_x509_attributes}
\alias{parse_x509_attributes}
\title{Parse X.509/X.500 Attribute Strings into a Named List}
\usage{
parse_x509_attributes(x, as = c("list", "df", "char"))
}
\arguments{
\item{x}{X.509 attribute string}

\item{as}{return type. \code{list} (named list), \code{char} (named character vector),
\code{df} (data frame).}
}
\value{
list, data frame or character vector depending on \code{as}
}
\description{
Parse X.509/X.500 Attribute Strings into a Named List
}
\examples{
x5_attr <- "CN=Sample Cert, OU=R&D, O=Company Ltd., L=Dublin 4, S=Dublin, C=IE"
parse_x509_attributes(x5_attr)
parse_x509_attributes(x5_attr, "df")
parse_x509_attributes(x5_attr, "char")
}
